Adding unit tests for execution_find_results.py & fixing syntax/spelling/names
